+++ CHANGELOG.md
@@ -16,6 +16,7 @@ The format is based on Keep a Changelog (https://keepachangelog.com), and this p
 - Add provider's interface unit tests for video package.
 - Add Mode interface and ModeInfo struct unit tests for video package.
 - Add EbitenDevice interface unit tests for video package.
+- Add Device interface unit tests for video package.
 
 ## [Unreleased] - 2026-01-29
 ### Added